equilibrium solvers moved to separate file "_solver.py"
